refactor(#94): layoutContext 내 mainContainer 컴포넌트 분리
